refactor(manage): tighten workspace governance
