+++ src/features/analytics-observability/metrics.ts
@@ -150,182 +150,182 @@ export const WORKBENCH_ANALYTICS_UI_OBSERVED_SURFACES = [
     operation: "reporting.archive-document.metadata",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "mandate-command-center",
     operation: "dpm.command-center.summary",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "mandate-command-center-monitoring",
     operation: "dpm.command-center.monitoring.run-once",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "mandate-command-center-exceptions",
     operation: "dpm.command-center.exceptions.list",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "mandate-command-center-exception-ai-summary",
     operation: "dpm.command-center.exceptions.ai-summary",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "mandate-command-center-mandate",
     operation: "dpm.command-center.mandate.by-portfolio",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "mandate-command-center-health",
     operation: "dpm.command-center.mandate.health",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "portfolio-memory",
     operation: "dpm.portfolio-memory.get",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "outcome-review-list",
     operation: "dpm.outcome-reviews.list",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "outcome-review-report-input",
     operation: "dpm.outcome-review.report-input",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "outcome-review-report-job",
     operation: "dpm.outcome-review.report-job.submit",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "outcome-review-ai-evidence",
     operation: "dpm.outcome-review.ai-evidence",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "outcome-review-ai-narrative",
     operation: "dpm.outcome-review.ai-narrative",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "construction-alternatives",
     operation: "dpm.construction.alternatives.generate",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "construction-alternative-set",
     operation: "dpm.construction.alternative-set.get",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "construction-selection",
     operation: "dpm.construction.alternative.select",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "proof-pack-generate",
     operation: "dpm.proof-pack.generate",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "proof-pack-detail",
     operation: "dpm.proof-pack.get",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "proof-pack-markdown",
     operation: "dpm.proof-pack.markdown",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "proof-pack-report-input",
     operation: "dpm.proof-pack.report-input",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "proof-pack-ai-evidence",
     operation: "dpm.proof-pack.ai-evidence",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "proof-pack-ai-pm-memo",
     operation: "dpm.proof-pack.ai-pm-memo",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "wave-list",
     operation: "dpm.waves.list",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "wave-preview",
     operation: "dpm.waves.preview",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "wave-create",
     operation: "dpm.waves.create",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "wave-detail",
     operation: "dpm.waves.get",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "wave-items",
     operation: "dpm.waves.items",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "wave-source-check",
     operation: "dpm.waves.source-check",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "wave-simulate",
     operation: "dpm.waves.simulate",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "wave-approve",
     operation: "dpm.waves.approve",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "wave-stage",
     operation: "dpm.waves.stage",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "wave-handoff",
     operation: "dpm.waves.handoff",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "wave-proof-pack",
     operation: "dpm.waves.proof-pack",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "wave-supportability",
     operation: "dpm.waves.supportability",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "wave-report-input",
     operation: "dpm.waves.report-input",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "wave-ai-pm-memo",
     operation: "dpm.waves.ai-pm-memo",
   },
   {
-    route: "workbench.dpm-command-center",
+    route: "workbench.manage",
     panel: "wave-operations-handoff-summary",
     operation: "dpm.waves.operations-handoff-summary",
   },
