+++ tests/unit/analytics-observability-metrics.test.ts
@@ -300,142 +300,142 @@ describe("analytics UI observability metrics", () => {
         "reporting.archive-document.metadata",
       ],
       [
-        "workbench.dpm-command-center",
+        "workbench.manage",
         "mandate-command-center",
         "dpm.command-center.summary",
       ],
       [
-        "workbench.dpm-command-center",
+        "workbench.manage",
         "mandate-command-center-monitoring",
         "dpm.command-center.monitoring.run-once",
       ],
       [
-        "workbench.dpm-command-center",
+        "workbench.manage",
         "mandate-command-center-exceptions",
         "dpm.command-center.exceptions.list",
       ],
       [
-        "workbench.dpm-command-center",
+        "workbench.manage",
         "mandate-command-center-exception-ai-summary",
         "dpm.command-center.exceptions.ai-summary",
       ],
       [
-        "workbench.dpm-command-center",
+        "workbench.manage",
         "mandate-command-center-mandate",
         "dpm.command-center.mandate.by-portfolio",
       ],
       [
-        "workbench.dpm-command-center",
+        "workbench.manage",
         "mandate-command-center-health",
         "dpm.command-center.mandate.health",
       ],
       [
-        "workbench.dpm-command-center",
+        "workbench.manage",
         "portfolio-memory",
         "dpm.portfolio-memory.get",
       ],
       [
-        "workbench.dpm-command-center",
+        "workbench.manage",
         "outcome-review-list",
         "dpm.outcome-reviews.list",
       ],
       [
-        "workbench.dpm-command-center",
+        "workbench.manage",
         "outcome-review-report-input",
         "dpm.outcome-review.report-input",
       ],
       [
-        "workbench.dpm-command-center",
+        "workbench.manage",
         "outcome-review-report-job",
         "dpm.outcome-review.report-job.submit",
       ],
       [
-        "workbench.dpm-command-center",
+        "workbench.manage",
         "outcome-review-ai-evidence",
         "dpm.outcome-review.ai-evidence",
       ],
       [
-        "workbench.dpm-command-center",
+        "workbench.manage",
         "outcome-review-ai-narrative",
         "dpm.outcome-review.ai-narrative",
       ],
       [
-        "workbench.dpm-command-center",
+        "workbench.manage",
         "construction-alternatives",
         "dpm.construction.alternatives.generate",
       ],
       [
-        "workbench.dpm-command-center",
+        "workbench.manage",
         "construction-alternative-set",
         "dpm.construction.alternative-set.get",
       ],
       [
-        "workbench.dpm-command-center",
+        "workbench.manage",
         "construction-selection",
         "dpm.construction.alternative.select",
       ],
       [
-        "workbench.dpm-command-center",
+        "workbench.manage",
         "proof-pack-generate",
         "dpm.proof-pack.generate",
       ],
       [
-        "workbench.dpm-command-center",
+        "workbench.manage",
         "proof-pack-detail",
         "dpm.proof-pack.get",
       ],
       [
-        "workbench.dpm-command-center",
+        "workbench.manage",
         "proof-pack-markdown",
         "dpm.proof-pack.markdown",
       ],
       [
-        "workbench.dpm-command-center",
+        "workbench.manage",
         "proof-pack-report-input",
         "dpm.proof-pack.report-input",
       ],
       [
-        "workbench.dpm-command-center",
+        "workbench.manage",
         "proof-pack-ai-evidence",
         "dpm.proof-pack.ai-evidence",
       ],
       [
-        "workbench.dpm-command-center",
+        "workbench.manage",
         "proof-pack-ai-pm-memo",
         "dpm.proof-pack.ai-pm-memo",
       ],
-      ["workbench.dpm-command-center", "wave-list", "dpm.waves.list"],
-      ["workbench.dpm-command-center", "wave-preview", "dpm.waves.preview"],
-      ["workbench.dpm-command-center", "wave-create", "dpm.waves.create"],
-      ["workbench.dpm-command-center", "wave-detail", "dpm.waves.get"],
-      ["workbench.dpm-command-center", "wave-items", "dpm.waves.items"],
+      ["workbench.manage", "wave-list", "dpm.waves.list"],
+      ["workbench.manage", "wave-preview", "dpm.waves.preview"],
+      ["workbench.manage", "wave-create", "dpm.waves.create"],
+      ["workbench.manage", "wave-detail", "dpm.waves.get"],
+      ["workbench.manage", "wave-items", "dpm.waves.items"],
       [
-        "workbench.dpm-command-center",
+        "workbench.manage",
         "wave-source-check",
         "dpm.waves.source-check",
       ],
-      ["workbench.dpm-command-center", "wave-simulate", "dpm.waves.simulate"],
-      ["workbench.dpm-command-center", "wave-approve", "dpm.waves.approve"],
-      ["workbench.dpm-command-center", "wave-stage", "dpm.waves.stage"],
-      ["workbench.dpm-command-center", "wave-handoff", "dpm.waves.handoff"],
-      ["workbench.dpm-command-center", "wave-proof-pack", "dpm.waves.proof-pack"],
+      ["workbench.manage", "wave-simulate", "dpm.waves.simulate"],
+      ["workbench.manage", "wave-approve", "dpm.waves.approve"],
+      ["workbench.manage", "wave-stage", "dpm.waves.stage"],
+      ["workbench.manage", "wave-handoff", "dpm.waves.handoff"],
+      ["workbench.manage", "wave-proof-pack", "dpm.waves.proof-pack"],
       [
-        "workbench.dpm-command-center",
+        "workbench.manage",
         "wave-supportability",
         "dpm.waves.supportability",
       ],
       [
-        "workbench.dpm-command-center",
+        "workbench.manage",
         "wave-report-input",
         "dpm.waves.report-input",
       ],
       [
-        "workbench.dpm-command-center",
+        "workbench.manage",
         "wave-ai-pm-memo",
         "dpm.waves.ai-pm-memo",
       ],
       [
-        "workbench.dpm-command-center",
+        "workbench.manage",
         "wave-operations-handoff-summary",
         "dpm.waves.operations-handoff-summary",
       ],
