+++ wiki/Home.md
@@ -16,9 +16,10 @@
 - primary product client for Lotus
 - Portfolio and Performance are the most mature active front-office surfaces
 - `/data-products` provides gateway-backed domain-product discovery and live trust posture
-- `/workbench/{portfolioId}` includes Gateway-backed DPM mandate command-center, rebalance-wave,
-  construction alternatives, proof-pack evidence with governed PM memo requests, and outcome-review panels for manage-owned
-  operating evidence
+- `/workbench/{portfolioId}` is the Manage workspace, with focused Gateway-backed sub-surfaces for
+  mandate command center, rebalance waves, construction alternatives, portfolio memory,
+  outcome reviews, proof-pack evidence with governed PM memo requests, and manage-owned operating
+  evidence
 - recommendations and proposals remain compatibility routes, not the main supported shell apps
 - shell navigation currently treats `Proposal` and `Advisory` as disabled capability-gated entries
