+++ wiki/RFC-Index.md
@@ -14,8 +14,9 @@
   DPM mandate command-center product experience backed by Gateway RFC-0098. It now includes
   implemented RFC-0038 command-center cockpit rendering, RFC-0039 construction alternatives,
   RFC-0040 proof-pack evidence, RFC-0041 rebalance-wave command-center rendering, and RFC-0042
-  post-trade outcome-review rendering on `/workbench/{portfolioId}`. Workbench consumes Gateway
-  only and does not calculate mandate health, wave readiness, proof-pack hashes,
+  post-trade outcome-review rendering across the focused Manage workspace modes under
+  `/workbench/{portfolioId}`. Workbench consumes Gateway only and does not calculate mandate
+  health, wave readiness, proof-pack hashes,
   expected-versus-realized values, AI prompts, or external execution state locally.
 
 ## Local references
